feat(ppo-validate-config): fail fast
