+++ verl/trainer/ppo/utils.py
@@ -1,0 +1,51 @@
+import warnings
+from enum import Enum
+
+from omegaconf import DictConfig
+
+from verl.single_controller.base import Worker
+from verl.trainer.ppo.core_algos import AdvantageEstimator
+
+WorkerType = type[Worker]
+
+
+class Role(Enum):
+    """
+    To create more roles dynamically, you can subclass Role and add new members
+    """
+
+    Actor = 0
+    Rollout = 1
+    ActorRollout = 2
+    Critic = 3
+    RefPolicy = 4
+    RewardModel = 5
+    ActorRolloutRef = 6
+
+
+def need_reference_policy(
+    role_worker_mapping: dict[Role, WorkerType],
+) -> bool:
+    """Given a role worker mapping, do we need ref policy."""
+    return Role.RefPolicy in role_worker_mapping
+
+
+def need_reward_model(
+    role_worker_mapping: dict[Role, WorkerType],
+) -> bool:
+    """Given a role worker mapping, do we need reward model."""
+    return Role.RewardModel in role_worker_mapping
+
+
+def need_critic(config: DictConfig) -> bool:
+    """Given a config, do we need ref policy."""
+    if config.critic.enable is not None:
+        return bool(config.critic.enable)
+    elif config.algorithm.adv_estimator == AdvantageEstimator.GAE:
+        return True
+    else:
+        warnings.warn(
+            "Disabled critic as algorithm.adv_estimator != gae. If it is not intended, please set critic.enable=True",
+            stacklevel=2,
+        )
+        return False
